🚚 #35 refactor: deleteContent 서버 액션 위치 이동
- 라우트를 점유하지 않는 액션은 app/actions 하위에 둔다
