+++ app/actions/deleteContent.ts
